% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethTF.R
\docType{data}
\name{clinical}
\alias{clinical}
\title{TCGA-COAD clinical matrix for 38 samples
Retrieved from GDC using TCGAbiolinks}
\format{
A matrix: 38 samples
}
\description{
TCGA-COAD clinical matrix for 38 samples
Retrieved from GDC using TCGAbiolinks
}
