% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethTF.R
\docType{package}
\name{coMethTF}
\alias{coMethTF}
\title{coMethTF: functional annotation of DMRs identified in epigenome-wide association studies}
\description{
To provide functional annotations for differentially methylated regions (DMRs)
and differentially methylated CpG sites (DMS), \code{coMethTF} performs
integrative analyses using matched DNA methylation and gene expression
along with Transcription Factor Binding Sites (TFBS) data.
coMethTF evaluates, prioritizes and annotates DNA methylation regions
(or sites) with high regulatory potential that works synergistically with
TFs to regulate target gene expressions, without any additional ChIP-seq data.
}
