% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDNAm.target.R
\name{corMetGene}
\alias{corMetGene}
\title{Evaluate correlation of region and gene using spearman test}
\usage{
corMetGene(links, met, exp, file.out, min.cor.fdr = 0.05, min.cor.estimate = 0)
}
\arguments{
\item{met}{DNA methylation matrix (rows are regions, columns samples). Samples should be in the
same order as gene expression.}

\item{exp}{Gene expression matrix (rows are genes, columns samples)
Samples should be in the same order as gene expression.}

\item{file.out}{A csv file name to save the results.}

\item{links.df}{A dataframe with the following collumns
region and gene}
}
\description{
Evaluate correlation of region and gene using spearman test
}
\examples{
regions.gr <- data.frame(
 chrom = c("chr22", "chr22", "chr22", "chr22", "chr22"),
  start = c("39377790", "50987294", "19746156", "42470063", "43817258"),
  end   = c("39377930", "50987527", "19746368", "42470223", "43817384"),
                         stringsAsFactors = FALSE)  \%>\%
                         makeGRangesFromDataFrame
map <- getDNAm.target(regions.gr = regions.gr ,genome = "hg19")
map <- unite(map,col = "regionID",c("region_chrom", "region_start", "region_end" ))
links <- tibble::tibble(regionID = map$regionID, geneID = map$ensembl_gene_id)
met <- matrix(rep(0,length(links$regionID) * 4),
              nrow = length(links$regionID),
              dimnames = list(c(links$regionID),c(paste0("S",c(1:4)))))

exp <- matrix(rep(0,length(links$geneID) * 4),
              nrow = length(links$geneID),
              dimnames = list(c(links$geneID),c(paste0("S",c(1:4)))))
corMetGene(links, met, exp)
}
