% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethTF.R
\docType{data}
\name{dna.met.chr21}
\alias{dna.met.chr21}
\title{A DNA methylation matrix for 50 samples (only chrmossome 21)}
\format{A matrix: 50 samples}
\description{
A DNA methylation matrix for 50 samples (only chrmossome 21)
}
