% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_quantile.R
\name{filter_dnam_by_quant_diff}
\alias{filter_dnam_by_quant_diff}
\title{Select regions with variations in DNA methylation
levels above a threshold}
\usage{
filter_dnam_by_quant_diff(dnam, diff.mean.th = 0.2, cores = 1)
}
\arguments{
\item{dnam}{DNA methylation matrix or SumarizedExperiment object}

\item{diff.mean.th}{Threshold for difference in mean DNAm levels for samples in Q4 and Q1}

\item{cores}{Number of CPU cores to be used in the analysis. Default: 1}
}
\value{
A subset of the original matrix only with the rows passing the filter threshold.
}
\description{
For each region, compares the mean DNA methylation (DNAm) levels
in samples with high DNAm  (Q4) vs. low DNAm (Q1) and requires
the difference to be above a threshold.
}
\examples{
data("dna.met.chr21")
dna.met.chr21.filtered <- filter_dnam_by_quant_diff(
  dna.met.chr21
)
}
