% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_quantile.R
\name{filter_genes_by_quantile_mean_fold_change}
\alias{filter_genes_by_quantile_mean_fold_change}
\title{Filter regions with high DNA methylation variance}
\usage{
filter_genes_by_quantile_mean_fold_change(exp, fold.change = 1.5, cores = 1)
}
\arguments{
\item{fold.change}{Mean fold change threshold between samples in Q4 and Q1. Default 1.5}

\item{cores}{Number of CPU cores to be used in the analysis. Default: 1}

\item{dnam}{DNA methylation matrix}
}
\description{
For each region, compares the mean DNA methylation of samples in high (Q4) vs. low (Q1)
and requires difference to be above a certain threshold.
}
\examples{
data("gene.exp.chr21")
gene.exp.chr21.filtered <- filter_genes_by_quantile_mean_fold_change(gene.exp.chr21)
}
