% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_regions.R
\name{filter_regions_by_mean_quantile_difference}
\alias{filter_regions_by_mean_quantile_difference}
\title{Filter regions with high DNA methylation variance}
\usage{
filter_regions_by_mean_quantile_difference(dnam, diff.mean.th = 0.2)
}
\arguments{
\item{dnam}{DNA methylation matrix}

\item{diff.mean.th}{Diff. mean Threshold between samples in Q4 and Q1}
}
\description{
For each region, compares the mean DNA methylation of samples in high (Q4) vs. low (Q1)
and requires difference to be above a certain threshold.
}
\examples{
data("dna.met.chr21")
dna.met.chr21.filtered <- filter_regions_by_mean_quantile_difference(dna.met.chr21)
}
