% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethReg.R
\docType{data}
\name{gene.exp.chr21.log2}
\alias{gene.exp.chr21.log2}
\title{TCGA-COAD gene expression matrix (log2 (FPKM-UQ + 1))
for 38 samples (only chromosome 21) retrieved from GDC using TCGAbiolinks}
\format{
A log2 (FPKM-UQ + 1) gene expression matrix with 38 samples,
includes Ensembl IDs in the rows and TCGA sample identifiers in the columns
}
\description{
TCGA-COAD gene expression matrix (log2 (FPKM-UQ + 1))
for 38 samples (only chromosome 21) retrieved from GDC using TCGAbiolinks
}
