% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_targets_cistrome.R
\name{get_cistrome_studies}
\alias{get_cistrome_studies}
\title{Get list of cistrome studies}
\usage{
get_cistrome_studies(conn = NULL)
}
\description{
Get list of cistrome studies
}
\examples{
get_cistrome_studies()
}
