% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hocomoco.R
\name{get_region_tf}
\alias{get_region_tf}
\title{Get human TF list for a region using HOCOMOCO prediction}
\usage{
get_region_tf(
  region,
  genome = c("hg19", "hg38"),
  arrayType = c("450k", "EPIC"),
  classification = c("subfamily", "family")
)
}
\value{
A dataframe with region, a TF name and TF gene ensembl ID
}
\description{
This function uses a pre-computed dataset for EPIC and HM450 Array that was created as follow:
For each HOCOMOCO human TF, the motif was search around the probe (+-250bp), and a binary matrix was created, with 1
if the motif was found, 0 if not. This function uses this pre-computed dataset to extend the probes to region, using the overlap probes
overlapping the regions as follows: for each region, get the probes within it and a motif will be one if at least one of the
overlapping probes has the motif (value 1 in the original dataset).
Then for each TF motifs found within the region, we select the TFs within the same TF family/subfamily since they
have similar binding motifs.
}
\examples{
\dontrun{
 regions.names <- c("chr22:18267969-18268249","chr23:18267969-18268249")
 get_region_tf("chr22:18267969-18268249",
                 genome = "hg19",
                 arrayType = "450k",
                 classification = "subfamily")
}
}
