% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_targets_cistrome.R
\name{get_tf_targets_cistrome}
\alias{get_tf_targets_cistrome}
\title{Access TF targets from Cistrome cancer database}
\usage{
get_tf_targets_cistrome(tcga.study, minCor = 0.2)
}
\arguments{
\item{tcga.study}{A TCGA study from the database
(i.e. "ACC","BLCA", "BRCA_1", "BRCA_2", "CESC", "COAD_READ*")
Please check function for a complete list: get_cistrome_studies()}

\item{minCor}{Cistrome minimum correlation between TF and target gene}
}
\value{
A dataframe with TF, target and correlation
}
\description{
Access cistrome TF cancer database
which is a comprehensive resource for predicted transcription factor (TF)
targets and enhancer profiles in cancers.
For more information: http://cistrome.org/CistromeCancer/
Note that it will take some time for this function to download
the cistrome cancer database (size ~1GB).
}
\examples{
\dontrun{
 get_tf_targets_cistrome("ACC")
 get_tf_targets_cistrome("COAD_READ*")
}
}
