% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triplet.R
\name{get_triplet}
\alias{get_triplet}
\title{For a given dnam region maps candidate target gene and scan for TF binding}
\usage{
get_triplet(
  region,
  genome = c("hg38", "hg19"),
  target.method = c("closest.gene", "window", "nearest.genes"),
  target.window.size = 500 * 10^3,
  target.num.flanking.genes = 5,
  motif.search.window.size = 0,
  motif.search.p.cutoff = 1e-08,
  cores = 1
)
}
\arguments{
\item{region}{A Granges or a named vector with regions (i.e "chr21:100002-1004000")}

\item{genome}{Human genome of reference "hg38" or "hg19"}

\item{target.method}{How genes are mapped to regions: closest gene promoter to the region ("closest.gene"); or
genes within a window around the region ("window").}

\item{target.window.size}{When \code{method = "window"}, number of base pairs to extend the region (+- window.size/2).
Default is 500kbp (or +/- 250kbp, i.e. 250k bp from start or end of the region)}

\item{motif.search.window.size}{Integer value to extend the regions. For example, a value of 50 will
extend 25 bp upstream and 25 downstream the region. Default is no increase}

\item{motif.search.p.cutoff}{Motifmatcher pvalue cut-off. Default 1e-8.}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\description{
This function wraps two other functions:
get_region_target_gene and get_tf_in_region from the package.
This function will map a region to a target gene using two methods
one mapping to the closest gene, and another one mapping to any gene within
a given window.
To find TF binding to the region JASPAR2020 is used to look into the region.
}
\examples{
\dontrun{
data("dna.met.chr21")
dnam.regions <- map_probes_to_regions(dna.met.chr21)
triplet <- get_triplet(
   region = rownames(dnam.regions)[1:100],
   motif.search.window.size = 50
)
}
}
