% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanTF.R
\name{getHumanTFs}
\alias{getHumanTFs}
\title{Access human TF from Lambert et al 2018}
\usage{
getHumanTFs(tcga.study)
}
\value{
A dataframe with Human TF
}
\description{
Access human TF from Lambert et al 2018
}
\examples{
\dontrun{
 human.tfs <- getHumanTFs()
}
}
