% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome.R
\name{getTFtargetsCistrome}
\alias{getTFtargetsCistrome}
\title{Access TF targets from Cistrome cancer database}
\usage{
getTFtargetsCistrome(tcga.study, minCor = 0.2)
}
\arguments{
\item{tcga.study}{A TCGA study from the database
(i.e. "ACC","BLCA", "BRCA_1", "BRCA_2", "CESC", "COAD_READ*")
Please check function for a complete list: getCistromSudies()}
}
\value{
A dataframe with TF, target and correlation
}
\description{
Access cistrome TF cancer database
which is a comprehensive resource for predicted transcription factor (TF)
targets and enhancer profiles in cancers.
For more information: http://cistrome.org/CistromeCancer/
}
\examples{
\dontrun{
 getTFtargetsCistrome("ACC")
 getTFtargetsCistrome("COAD_READ*")
}
}
