% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_names_from_granges}
\alias{make_names_from_granges}
\title{Create region name from Granges}
\usage{
make_names_from_granges(region)
}
\description{
Given a GRanges returns region name such as chr22:18267969-18268249
}
\examples{
regions.names <- c("chr22:18267969-18268249","chr23:18267969-18268249")
regions.gr <- make_granges_from_names(regions.names)
make_names_from_granges(regions.gr)
}
