% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_se_from_dnam_regions}
\alias{make_se_from_dnam_regions}
\title{Transform DNA methylation array to a summarized Experiment object}
\usage{
make_se_from_dnam_regions(met)
}
\arguments{
\item{met}{DNA methylation matrix with beta-values or m-values as data,
row as cpgs and column as samples}
}
\description{
Transform DNA methylation array to a summarized Experiment object
}
\examples{
dna.met.chr21 <- get(data("dna.met.chr21"))
dna.met.chr21.regions <- map_probes_to_regions(dna.met.chr21)
dnam.se <- make_se_from_dnam_regions(dna.met.chr21.regions)
}
