% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_se_from_gene_matrix}
\alias{make_se_from_gene_matrix}
\title{Transform gene expression matrix into a Summarized Experiment object}
\usage{
make_se_from_gene_matrix(exp, genome = c("hg38", "hg19"))
}
\arguments{
\item{exp}{Gene expression matrix with gene expression counts,
row as ENSG gene IDS and column as samples}

\item{genome}{Human genome of reference: hg38 or hg19}
}
\value{
A summarized Experiment object
}
\description{
Transform gene expression matrix into a Summarized Experiment object
}
\examples{
gene.exp.chr21.log2 <- get(data("gene.exp.chr21.log2"))
gene.exp.chr21.log2.se <- make_se_from_gene_matrix(gene.exp.chr21.log2)
}
