% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hocomoco.R
\name{makeGrangesFromNames}
\alias{makeGrangesFromNames}
\title{Create granges from name}
\usage{
makeGrangesFromNames(names)
}
\description{
Create granges from name
}
\examples{
regions.names <- c("chr22:18267969-18268249","chr23:18267969-18268249")
regions.gr <- makeGrangesFromNames(regions.names)
}
