% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_probes_to_regions}
\alias{map_probes_to_regions}
\title{Change probes names to region names}
\usage{
map_probes_to_regions(
  dnam,
  genome = c("hg38", "hg19"),
  arrayType = c("450k", "EPIC")
)
}
\arguments{
\item{dnam}{A DNA methylation matrix}

\item{genome}{Human genome of reference hg38 or hg19}

\item{arrayType}{DNA methylation array type (450k or EPIC)}
}
\description{
Given a DNA methylation matrix with probes as row names,
map probes to genomic regions
}
\examples{
data(dna.met.chr21)
dna.met.chr21.with.region.name <- map_probes_to_regions(dna.met.chr21)
}
