% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hocomoco.R
\name{mapTFBSHocomoco}
\alias{mapTFBSHocomoco}
\title{Get human TF list for region using hocomoco prediction}
\usage{
mapTFBSHocomoco(
  region,
  genome = c("hg19", "hg38"),
  arrayType = c("450k", "EPIC"),
  classification = c("subfamily", "family")
)
}
\value{
A dataframe with TF, target and correlation
}
\description{
Get list of TF mapping to regions using
TF motifs to regions
}
\examples{
\dontrun{
 regions.names <- c("chr22:18267969-18268249","chr23:18267969-18268249")
 mapTFBSHocomoco("chr22:18267969-18268249",
                 genome = "hg19",
                 arrayType = "450k",
                 classification = "subfamily")
}
}
