% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethReg-package.R
\docType{package}
\name{MethReg-package}
\alias{MethReg}
\alias{MethReg-package}
\title{MethReg: functional annotation of DMRs identified in epigenome-wide
association studies}
\description{
To provide functional annotations for
differentially methylated regions (DMRs)
and differentially methylated CpG sites (DMS), \code{MethReg} performs
integrative analyses using matched DNA methylation and gene expression
along with Transcription Factor Binding Sites (TFBS) data.
MethReg evaluates, prioritizes and annotates DNA methylation regions
(or sites) with high regulatory potential that works synergistically with
TFs to regulate target gene expressions, without any
additional ChIP-seq data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/TransBioInfoLab/MethReg/issues/}
}

}
\author{
\strong{Maintainer}: Tiago Silva \email{tiagochst@gmail.com} (\href{https://orcid.org/0000-0003-1343-6850}{ORCID})

Authors:
\itemize{
  \item Lily Wang \email{lily.wangg@gmail.com}
}

}
\keyword{internal}
