% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionModel.R
\name{plot_interaction_model}
\alias{plot_interaction_model}
\title{Plot interaction data}
\usage{
plot_interaction_model(triplet.results, dnam, exp)
}
\arguments{
\item{triplet.results}{Output from function interaction_model
with Region ID, TF  (column name: TF),  and target gene  (column name: target),
p-vallues and estimates of interaction}

\item{dnam}{DNA methylation matrix  (columns: samples same order as met, rows: regions/probes)}

\item{exp}{gene expression matrix (columns: samples same order as met, rows: genes)}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
Create several plots to show interaction data
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) ~ log2(TF) + DNAm + log2(TF) * DNAm}

To consider covariates, RNA can also be the residuals.
\deqn{log2(RNA target residuals) ~ log2(TF residual) + DNAm + log2(TF residual) * DNAm}
}
\examples{
triplet <- data.frame("regionID" = paste0("region_",1:9),
                      "TF" = paste0("ESNG",10),
                      "target" = paste0("ESNG",1:9))
dnam <- runif(200) \%>\% matrix(10)
colnames(dnam) <- paste0("Sample_",1:20)
rownames(dnam) <- paste0("region_",1:10)
exp <- rexp(200) \%>\% matrix(10)
colnames(exp) <- paste0("Sample_",1:20)
rownames(exp) <- paste0("ESNG",1:10)
results <- interaction_model(triplet, dnam, exp)
plots <- plot_interaction_model(results, dnam, exp)
}
