% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap2022.R
\name{readRemap2022}
\alias{readRemap2022}
\title{Access REMAP2022 non-redundant peaks}
\usage{
readRemap2022(cell_line, species = c("homo_sapiens", "mus_musculus"))
}
\arguments{
\item{cell_line}{filter peaks using cell line description field}

\item{species}{which species to access REMAP2022 non-redundant peaks.
Options are: homo_sapiens or mus_musculus}
}
\description{
Access REMAP2022 non-redundant peaks
}
