% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome.R
\name{registerCores}
\alias{registerCores}
\title{register cores}
\usage{
registerCores(cores)
}
\arguments{
\item{cores}{A interger which defines the number of cores to be used in parallel}
}
\description{
register cores
}
