% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_analysis.R
\name{stratified_model}
\alias{stratified_model}
\title{Fits linear model to triplet data (Target, TF, DNAm) separately for
samples with DNAm high or low groups.}
\usage{
stratified_model(triplet, dnam, exp, cores = 1)
}
\arguments{
\item{triplet}{Data frame with columns for DNA methylation region (regionID), TF  (TF), and target gene  (target)}

\item{dnam}{DNA methylation matrix  (columns: samples in the same order as \code{exp} matrix, rows: regions/probes)}

\item{exp}{A log2 (gene expression + 1) matrix (columns: samples in the same order as \code{dnam} matrix,
rows: genes represented by ensembl IDs (e.g. ENSG00000239415))}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
Should be used only for triplet data with significant
 \code{TF*DNAm} interaction from fitting models in \code{interaction_model}.
These models can be used to examine how TF activities differ in
samples with high DNAm or low DNAm values.
}
\details{
This function fits linear model
\code{log2(RNA target) ~ log2(TF)}

to samples with higest DNAm values (top 25 percent) and lowest DNAm values (bottom 25 percent), separately.

To account for confounding effects from covariate variables, first use the \code{get_residuals} function to obtain
RNA residual values which have covariate effects removed, then fit interaction model. Note that no
log2 transformation is needed when \code{interaction_model} is applied to residuals data.
}
\examples{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21")
triplet <- data.frame("regionID" = rownames(dna.met.chr21)[1:10],
                      "TF" = rownames(gene.exp.chr21)[11:20],
                      "target" = rownames(gene.exp.chr21)[1:10])
results <- stratified_model(triplet, dna.met.chr21, gene.exp.chr21)
}
