% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethReg.R
\docType{data}
\name{clinical}
\alias{clinical}
\title{TCGA-COAD clinical matrix for 38 samples
retrieved from GDC using \link[TCGAbiolinks]{TCGAbiolinks}}
\format{
A matrix: 38 samples (rows) and variables (columns)
patient, sample, gender and sample_type
}
\usage{
clinical
}
\description{
TCGA-COAD clinical matrix for 38 samples
retrieved from GDC using \link[TCGAbiolinks]{TCGAbiolinks}
}
\keyword{datasets}
