% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_region_dnam_target_gene.R
\name{cor_region_dnam_target_gene}
\alias{cor_region_dnam_target_gene}
\title{Evaluate correlation of DNA methylation region and target gene expression}
\usage{
cor_region_dnam_target_gene(
  links,
  met,
  exp,
  min.cor.pval = 0.05,
  min.cor.estimate = 0,
  file.out,
  cores = 1
)
}
\arguments{
\item{links}{A dataframe with the following columns: regionID (DNA methylation) and target (target gene)}

\item{met}{DNA methylation matrix (rows are regions and columns are samples). Samples should be in the
same order as gene expression.}

\item{exp}{Gene expression matrix (rows are genes, columns are samples) log2-normalized (log2(exp + 1)).
Samples should be in the same order as the DNA methylation matrix.}

\item{min.cor.pval}{Filter of significant correlations (default: 0.05)}

\item{min.cor.estimate}{Filter of significant correlations (default: not applied)}

\item{file.out}{If provided, name of a csv file which will be used to save the results.}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\description{
Evaluate correlation of the DNA methylation region and target gene expression
using spearman correlation test
}
\examples{
library(dplyr)
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
regions.gr <- make_granges_from_names(rownames(dna.met.chr21))
data("gene.exp.chr21")

# Map example region to closest gene
links <- get_region_target_gene(regions.gr = regions.gr, genome = "hg19", method = "closest.gene")

# Samples in met and exp datasets should be in the same order.
identical (colnames (dna.met.chr21), colnames(gene.exp.chr21))

# Correalted DNAm and gene expression, display only significant associations
results <- cor_region_dnam_target_gene(links = links, met = dna.met.chr21, exp = gene.exp.chr21)

# display all associations
results.all <- cor_region_dnam_target_gene(
   links = links,
   met = dna.met.chr21,
   exp = gene.exp.chr21,
   min.cor.pval = 1)
}
