% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_region_dnam_target_gene.R
\name{cor_region_dnam_target_gene}
\alias{cor_region_dnam_target_gene}
\title{Evaluate correlation of DNA methylation region and target gene expression}
\usage{
cor_region_dnam_target_gene(
  links,
  dnam,
  exp,
  filter.results = TRUE,
  min.cor.pval = 0.05,
  min.cor.estimate = 0,
  file.out,
  cores = 1
)
}
\arguments{
\item{links}{A dataframe with the following columns: regionID (DNA methylation) and target (target gene)}

\item{dnam}{DNA methylation matrix (rows are regions and columns are samples). Samples should be in the
same order as gene expression.}

\item{exp}{Gene expression matrix (rows are genes, columns are samples) log2-normalized (log2(exp + 1)).
Samples should be in the same order as the DNA methylation matrix.}

\item{filter.results}{Filter results using min.cor.pval and min.cor.estimate thresholds}

\item{min.cor.pval}{Filter of significant correlations (default: 0.05)}

\item{min.cor.estimate}{Filter of significant correlations (default: not applied)}

\item{file.out}{If provided, name of a csv file which will be used to save the results.}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\value{
A data frame with the following information: regionID, target geene, correlation pvalue and estimate between
DNA methylation and target gene expression, FDR corrected p-values.
}
\description{
This function evaluate the correlation of the DNA methylation region and target gene expression
using spearman rank correlation test.  Note that genes with RNA expression equal to 0 for more than 25% of the samples
will not be evaluated.
}
\examples{
dnam <- t(matrix(sort(c(runif(20))), ncol = 1))
rownames(dnam) <- c("chr3:203727581-203728580")
colnames(dnam) <- paste0("Samples",1:20)
exp <- dnam
rownames(exp) <- c("ENSG00000232886")
colnames(exp) <- paste0("Samples",1:20)

links <- data.frame(
   "regionID" =  c("chr3:203727581-203728580"),
   "target" = "ENSG00000232886"
)

# Correalted DNAm and gene expression, display only significant associations
results.cor.pos <- cor_region_dnam_target_gene(
   links = links,
   dnam = dnam,
   exp = exp,
   filter.results = FALSE
)
\dontrun{
# Load data
data("gene.exp.chr21")
data("dna.met.chr21")
dna.met.chr21 <- make_se_from_dnam_probes(dna.met.chr21)

# Map example region to closest gene
links <- get_region_target_gene(
  regions.gr = dna.met.chr21,
  genome = "hg19",
  method = "closest.gene"
)

# Samples in met and exp datasets should be in the same order.
identical (colnames (dna.met.chr21), colnames(gene.exp.chr21))

# Correalted DNAm and gene expression, display only significant associations
results <- cor_region_dnam_target_gene(
  links = links,
  dnam = dna.met.chr21,
  exp = gene.exp.chr21
)

# display all associations
results.all <- cor_region_dnam_target_gene(
   links = links,
   dnam = dna.met.chr21,
   exp = gene.exp.chr21,
   min.cor.pval = 1
)
}
}
