% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethReg.R
\docType{data}
\name{dna.met.chr21}
\alias{dna.met.chr21}
\title{TCGA-COAD DNA methylation matrix (beta-values) for 38 samples (only chr21)
retrieved from GDC using TCGAbiolinks.}
\format{
A beta-value matrix with 38 samples, includes CpG IDs in the rows and
TCGA sample identifiers in the columns
}
\description{
TCGA-COAD DNA methylation matrix (beta-values) for 38 samples (only chr21)
retrieved from GDC using TCGAbiolinks.
}
