% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethTF.R
\docType{data}
\name{dna.met.chr21}
\alias{dna.met.chr21}
\title{TCGA-COAD DNA methylation matrix for 38 samples (only chrmossome 21)
retrieved from GDC using TCGAbiolinks}
\format{
A matrix with 38 samples, includes CpG IDs in the rows and
TCGA sample identifiers in the columns
}
\description{
TCGA-COAD DNA methylation matrix for 38 samples (only chrmossome 21)
retrieved from GDC using TCGAbiolinks
}
