% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_results_to_table}
\alias{export_results_to_table}
\title{Format methreg results table and export to XLSX file}
\usage{
export_results_to_table(results, file = "MethReg_results.xlsx")
}
\value{
A summarized Experiment object
}
\description{
Format methreg results table and export to XLSX file
}
\examples{
gene.exp.chr21.log2 <- get(data("gene.exp.chr21.log2"))
gene.exp.chr21.log2.se <- make_exp_se(gene.exp.chr21.log2)
}
