% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMethTF.R
\docType{data}
\name{gene.exp.chr21}
\alias{gene.exp.chr21}
\title{A gene expression matrix for 50 samples (only chrmossome 21)}
\format{
A matrix: 50 samples
}
\description{
A gene expression matrix for 50 samples (only chrmossome 21)
}
