% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanTF.R
\name{get_human_tfs}
\alias{get_human_tfs}
\title{Access human TF from Lambert et al 2018}
\usage{
get_human_tfs()
}
\value{
A dataframe with Human TF
}
\description{
Access human TF from Lambert et al 2018
}
\examples{
\dontrun{
 human.tfs <- get_human_tfs()
}
}
