% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_promoter_avg.R
\name{get_promoter_avg}
\alias{get_promoter_avg}
\title{Summarize promoter methylation beta values by mean or median.}
\usage{
get_promoter_avg(dnam, genome, arrayType, cores = 1)
}
\value{
A list with a data frame of the new regions and merged ones,
and the calculated mean/median beta-value matrix
}
\description{
For multiple probes mapped to the same promoter region,
take the mean or median beta-values
}
\examples{
data("dna.met.chr21")
promoter.avg <- get_promoter_avg(dna.met.chr21, genome = "hg19", arrayType = "450k")
}
