% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDNAm.target.R
\name{get_region_target_gene}
\alias{get_region_target_gene}
\title{Mapping regions to gene}
\usage{
get_region_target_gene(
  regions.gr,
  genome = c("hg38", "hg19"),
  method = c("closest.gene", "window"),
  window.width = 5e+05
)
}
\arguments{
\item{regions.gr}{A Genomic Ranges objec GRanges}

\item{genome}{Human genome of reference "hg38" or "hg19"}

\item{method}{How to map regions to genes: closest gene ("closest.gene)
genes within a window around the region ("window").
around the genomic region input.}

\item{window.width}{Number of base pairs to extend the region (+-window.width/2). Default is 500kbp (+- 250kbp)}
}
\description{
To map a region to genes there are two options: 1) closest gene
2) map to all genes within a window around the region (default 500kbp/+- 250kbp around the region)
}
\examples{
library(GenomicRanges)
library(dplyr)

# Create example region
regions.gr <- data.frame(
              chrom = c("chr22", "chr22", "chr22", "chr22", "chr22"),
              start = c("39377790", "50987294", "19746156", "42470063", "43817258"),
              end   = c("39377930", "50987527", "19746368", "42470223", "43817384"),
              stringsAsFactors = FALSE)  \%>\%
     makeGRangesFromDataFrame

 # map to closest gene
 get_region_target_gene(regions.gr = regions.gr, genome = "hg19", method = "closest.gene")

 # map to all gene within region +- 250kbp
 get_region_target_gene(regions.gr = regions.gr, genome = "hg19", method = "window")
}
