% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{get_residuals}
\alias{get_residuals}
\title{Get Residuals}
\usage{
get_residuals(data.matrix, metadata)
}
\arguments{
\item{data.matrix}{A matrix with samples as columns and features (gene, probes)
as rows.}

\item{metadata}{A data frame with samples as rows and columns the covariates}
}
\value{
A residuals matrix with samples as columns and features (gene, probes)
}
\description{
Get Residuals using a data matrix and a metadata data frame as
follows: features ~ covariate1 + covariate2 ... + covariateN
where n is the index of the columns of the metadata provided
}
\examples{
data("gene.exp.chr21")
metadata <- data.frame(
    row.names = paste0("Sample",idx),
    sex = c(rep("Male",25),rep("Female",25)),
    age = sample(50:90, 50, replace = TRUE)
)
gene.exp.residuals <- get_residuals(gene.exp.chr21,metadata)
}
