% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Get studentized residual from regression model}
\usage{
get_residuals(
  data.matrix,
  metadata.samples = NULL,
  metadata.genes = NULL,
  cores = 1
)
}
\arguments{
\item{data.matrix}{A matrix with samples as columns and features (gene, probes)
as rows.}

\item{metadata.samples}{A data frame with samples as rows and columns the covariates.
Observation: No NA values are allowed, otherwise residual of the sample will be NA.}

\item{metadata.genes}{A data frame with genes (covariates) as rows and samples on columns.
For each evaluated gene the columns will be set as a single covariates.
This can be used to correct each gene for copy number alterations.}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\value{
A residuals matrix with samples as columns and features (gene, probes)
}
\description{
Get Residuals using a data matrix and a sample metadata data frame as
follows: features ~ Sample_covariate1 + Sample_covariate2 ... + Sample_covariateN
where n is the index of the columns of the metadata provided
In case the user also provide a gene metadata (i.e copy number variation/alteration)
get Residuals using:
features ~ Sample_covariate1 + Sample_covariate2 ... + Sample_covariateN + gene_covariate
}
\examples{
data("gene.exp.chr21")
data("clinical")
metadata <- clinical[,c( "gender", "sample_type")]
cnv <- matrix(
   sample(x = c(-2,-1,0,1,2),
   size = ncol(gene.exp.chr21) * nrow(gene.exp.chr21),replace = TRUE),
   nrow = nrow(gene.exp.chr21),
   ncol = ncol(gene.exp.chr21)
)
rownames(cnv) <- rownames(gene.exp.chr21)
colnames(cnv) <- colnames(gene.exp.chr21)
gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21[1:3,],
   metadata.samples = metadata,
   metadata.genes = cnv
)
gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21[1:3,],
   metadata.samples = metadata,
   metadata.genes = cnv[1:2,]
)
gene.exp.residuals <- get_residuals(
   data.matrix = gene.exp.chr21[1:3,],
   metadata.samples = metadata
)
}
