% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{Get Residuals}
\usage{
get_residuals(data.matrix, metadata)
}
\arguments{
\item{data.matrix}{A matrix with samples as columns and features (gene, probes)
as rows.}

\item{metadata}{A data frame with samples as rows and columns the covariates}
}
\value{
A residuals matrix with samples as columns and features (gene, probes)
}
\description{
Get Residuals using a data matrix and a metadata data frame as
follows: features ~ covariate1 + covariate2 ... + covariateN
where n is the index of the columns of the metadata provided
}
\examples{
data("gene.exp.chr21")
data("clinical")
metadata <- clinical[,c( "gender", "sample_type")]
gene.exp.residuals <- get_residuals(gene.exp.chr21[1:3,], metadata)
}
