% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tf_in_region.R
\name{get_tf_in_region}
\alias{get_tf_in_region}
\title{Get human TFs for regions by scanning it with motifmatchr using JASPAR 2020 database}
\usage{
get_tf_in_region(
  region,
  window.size = 0,
  genome = c("hg19", "hg38"),
  p.cutoff = 1e-08,
  cores = 1
)
}
\arguments{
\item{region}{A vector of region names or GRanges object with the DNA methylation regions to be scanned for the motifs}

\item{window.size}{Integer value to extend the regions. For example, a value of 50 will
extend 25 bp upstream and 25 bp downstream the region.
The default is not to increase the scanned region.}

\item{genome}{Human genome of reference "hg38" or "hg19"}

\item{p.cutoff}{Motifmatcher p.cutoff. Default 1e-8.}

\item{cores}{Number of CPU cores to be used. Default 1.}
}
\value{
A data frame with the following information: regionID, TF symbol, TF ensembl ID
}
\description{
Given a genomic region, this function uses motifmatchr and JASPAR2020
to scan the region for 554 human transcription factors binding sites. There is also
an option (argument \code{window.size}) to extend the scanning region before performing the search, which
by default is 0 (do not extend)
}
\examples{
\dontrun{
 regions.names <- c("chr3:189631389-189632889","chr4:43162098-43163498")
 region.tf <- get_tf_in_region(
                 region = regions.names,
                 genome = "hg38"
 )

 regions.names <- c("chr1:79592-79592","chr4:43162198-43162198")
 regions.gr <- make_granges_from_names(regions.names)
 region.tf <- get_tf_in_region(
                 region = regions.gr,
                 window.size = 25,
                 genome = "hg38"
 )
}
}
