% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionModel.R
\name{interaction_model}
\alias{interaction_model}
\title{Calculate interaction linear model Target ~ TF + DNAm}
\usage{
interaction_model(met, exp, TF.target, DNAm.target)
}
\arguments{
\item{met}{DNA methylation matrix}

\item{exp}{gene EXpression matrix}

\item{TF.target}{Dataframe with TF and target gene}

\item{DNAm.target}{Dataframe with DNAm and target gene}
}
\value{
A dataframe with Region, TF, Estimates and Pvalue from linear model
}
\description{
Evaluate interaction of DNA methylation,
TF expression with target gene interaction using a linear model
log2(RNA target ) ~ log2(TF) + DNAm + log2(TF) * DNAm
}
\examples{
\dontrun{
 human.tfs <- get_human_tfs()
 TF.target <- get_tf_targets_cistrome("COAD_READ*")
 DNAm.target <- get_dnam_target_gene("hg38", method = "closest.gene")
}
}
