% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{interaction_model}
\alias{interaction_model}
\title{Calculate interaction linear model Target ~ TF + DNAm}
\usage{
interaction_model(triplet, dnam, exp)
}
\arguments{
\item{triplet}{Dataframe with region (column name: regionID),
TF  (column name: TF),  and target gene  (column name: target),}

\item{dnam}{DNA methylation matrix  (columns: samples same order as met, rows: regions/probes)}

\item{exp}{A log2 (gene expression + 1) matrix with samples as columns in the same order as met
and genes as rows represented by ensembl IDs ENSG00000239415)}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
Evaluate interaction of DNA methylation,
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) ~ log2(TF) + DNAm + log2(TF) * DNAm}

To consider covariates, RNA can also be the residuals.
\deqn{log2(RNA target residuals) ~ log2(TF residual) + DNAm + log2(TF residual) * DNAm}
}
\examples{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21")
triplet <- data.frame("regionID" = rownames(dna.met.chr21)[1:10],
                      "TF" = rownames(gene.exp.chr21)[11:20],
                      "target" = rownames(gene.exp.chr21)[1:10])
results <- interaction_model(triplet, dna.met.chr21, gene.exp.chr21)
}
