% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_granges_from_names}
\alias{make_granges_from_names}
\title{Create Granges from name}
\usage{
make_granges_from_names(names)
}
\description{
Given a region name such as chr22:18267969-18268249, we will create a Granges
object
}
\examples{
regions.names <- c("chr22:18267969-18268249","chr23:18267969-18268249")
regions.gr <- make_granges_from_names(regions.names)
}
