% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_se_from_dnam_probes}
\alias{make_se_from_dnam_probes}
\title{Transform DNA methylation array into a summarized Experiment object}
\usage{
make_se_from_dnam_probes(
  dnam,
  genome = c("hg38", "hg19"),
  arrayType = c("450k", "EPIC"),
  betaToM = FALSE
)
}
\arguments{
\item{dnam}{DNA methylation matrix with beta-values or m-values as data,
row as cpgs and column as samples}

\item{genome}{Human genome of reference: hg38 or hg19}

\item{arrayType}{DNA methylation array type (450k or EPIC)}

\item{betaToM}{indicates if converting methylation beta values to mvalues}
}
\value{
A summarized Experiment object with DNA methylation probes mapped to
genomic regions
}
\description{
Transform DNA methylation array into a summarized Experiment object
}
\examples{
\dontrun{
  dna.met.chr21 <- get(data("dna.met.chr21"))
  se <- make_se_from_dnam_probes(dna.met.chr21)
}
}
