% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_se_from_dnam_regions}
\alias{make_se_from_dnam_regions}
\title{Transform DNA methylation array into a summarized Experiment object}
\usage{
make_se_from_dnam_regions(dnam, betaToM = FALSE)
}
\arguments{
\item{dnam}{DNA methylation matrix with beta-values or m-values as data,
row as genomic regions and column as samples}

\item{betaToM}{indicates if converting methylation beta values to mvalues}
}
\value{
A summarized Experiment object
}
\description{
Transform DNA methylation array into a summarized Experiment object
}
\examples{
\dontrun {
  dna.met.chr21 <- get(data("dna.met.chr21"))
  dna.met.chr21.regions <- MethReg:::map_probes_to_regions(dna.met.chr21)
  dnam.se <- make_se_from_dnam_regions(dna.met.chr21.regions)
}
}
