% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractionModel.R
\name{plot_interaction_model}
\alias{plot_interaction_model}
\title{Plot interaction data}
\usage{
plot_interaction_model(triplet.results, dnam, exp, metadata)
}
\arguments{
\item{triplet.results}{Output from function interaction_model
with Region ID, TF  (column name: TF),  and target gene  (column name: target),
p-vallues and estimates of interaction}

\item{dnam}{DNA methylation matrix  (columns: samples same order as met, rows: regions/probes)}

\item{exp}{gene expression matrix (columns: samples same order as met, rows: genes)}

\item{metadata}{A data frame with samples as rownames and one columns that will be used to
color the samples}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
Create several plots to show interaction data
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) ~ log2(TF) + DNAm + log2(TF) * DNAm}

To consider covariates, RNA can also be the residuals.
\deqn{log2(RNA target residuals) ~ log2(TF residual) + DNAm + log2(TF residual) * DNAm}
}
\examples{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21")
triplet <- data.frame("regionID" = rownames(dna.met.chr21)[1:10],
                      "TF" = rownames(gene.exp.chr21)[11:20],
                      "target" = rownames(gene.exp.chr21)[1:10])
results <- interaction_model(triplet, dna.met.chr21, gene.exp.chr21)
metadata <- data.frame(row.names = colnames(dna.met.chr21),group = c(rep("Tumor",25),rep("Normal",25)))
plots <- plot_interaction_model(results[1,], dna.met.chr21, gene.exp.chr21,metadata)
}
