% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_model.R
\name{plot_interaction_model}
\alias{plot_interaction_model}
\title{Plot interaction model results}
\usage{
plot_interaction_model(
  triplet.results,
  dnam,
  exp,
  metadata,
  tf.activity.es = NULL,
  tf.dnam.classifier.pval.thld = 0.001,
  label.dnam = "beta-value",
  label.exp = "expression",
  genome = "hg38"
)
}
\arguments{
\item{triplet.results}{Output from function interaction_model
with Region ID, TF  (column name: TF),  and target gene  (column name: target),
p-values and estimates of interaction}

\item{dnam}{DNA methylation matrix or SummarizedExperiment object
(columns: samples same order as met, rows: regions/probes)}

\item{exp}{gene expression matrix or a SummarizedExperiment object
(columns: samples same order as met, rows: genes)}

\item{metadata}{A data frame with samples as rownames and one columns that will be used to
color the samples}

\item{tf.activity.es}{A matrix with normalized enrichment scores for each TF across all samples
to be used in linear models instead of TF gene expression.}

\item{tf.dnam.classifier.pval.thld}{P-value threshold to consider
a linear model significant
of not. Default 0.001. This will be used to classify the TF role and DNAm
effect.}

\item{label.dnam}{Used for label text. Option "beta-value" and "residuals"}

\item{label.exp}{Used for label text. Option "expression" and "residuals"}

\item{genome}{Genome of reference to be added to the plot as text}
}
\value{
A ggplot object, includes a table with results from fitting interaction model,
and the the following scatter plots: 1) TF vs DNAm, 2) Target vs DNAm,
3) Target vs TF, 4) Target vs TF for samples in Q1 and Q4 for DNA methylation,
5) Target vs DNAm for samples in Q1 and Q4 for the TF
}
\description{
Create several plots to show interaction data
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) = log2(TF) + DNAm + log2(TF) * DNAm}

To consider covariates, RNA can also be the residuals.
\deqn{log2(RNA target residuals) = log2(TF residual) + DNAm + log2(TF residual) * DNAm}
}
\examples{
library(dplyr)
dnam <- runif(20, min = 0,max = 1) \%>\% sort \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(dnam) <- c("chr3:203727581-203728580")
colnames(dnam) <- paste0("Samples",1:20)

exp.target <-  runif(20,min = 0,max = 10) \%>\% sort \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(exp.target) <- c("ENSG00000232886")
colnames(exp.target) <- paste0("Samples",1:20)

exp.tf <- runif(20,min = 0,max = 10) \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(exp.tf) <- c("ENSG00000101412")
colnames(exp.tf) <- paste0("Samples",1:20)

exp <- rbind(exp.tf, exp.target)

triplet <- data.frame(
   "regionID" =  c("chr3:203727581-203728580"),
   "target" = "ENSG00000232886",
   "TF" = "ENSG00000101412"
)

results <- interaction_model(
  triplet = triplet,
  dnam = dnam,
  exp = exp,
  fdr = FALSE,
  filter.correlated.tf.exp.dna = FALSE
)
plots <- plot_interaction_model(
    triplet.results = results,
    dnam = dnam,
    exp = exp
)
}
