% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratified_model.R
\name{plot_stratified_model}
\alias{plot_stratified_model}
\title{Plot stratifed model results}
\usage{
plot_stratified_model(triplet.results, dnam, exp, metadata)
}
\arguments{
\item{triplet.results}{Output from function stratified_model
with Region ID, TF  (column name: TF),  and target gene  (column name: target),
p-values and estimates of interaction}

\item{dnam}{DNA methylation matrix  (columns: samples same order as met, rows: regions/probes)}

\item{exp}{gene expression matrix (columns: samples same order as met, rows: genes)}

\item{metadata}{A data frame with samples as rownames and one columns that will be used to
color the samples}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
Create several plots to show interaction data
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) ~ log2(TF)}
to samples with higest DNAm values (top 25 percent) and lowest DNAm values (bottom 25 percent), separately.
}
\examples{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21")
triplet <- data.frame("regionID" = rownames(dna.met.chr21)[1:5],
                      "TF" = rownames(gene.exp.chr21)[11:15],
                      "target" = rownames(gene.exp.chr21)[1:5])
results <- stratified_model(triplet, dna.met.chr21, gene.exp.chr21)
plots <- plot_stratified_model(results[1,], dna.met.chr21, gene.exp.chr21)
# Adding color to samples
metadata <- clinical[,"sample_type",drop = FALSE]
plots <- plot_stratified_model(results[1,], dna.met.chr21, gene.exp.chr21,metadata)
}
