% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratified_model.R
\name{plot_stratified_model}
\alias{plot_stratified_model}
\title{Plot stratified model results}
\usage{
plot_stratified_model(triplet.results, dnam, exp, metadata)
}
\arguments{
\item{triplet.results}{Output from function stratified_model
with Region ID, TF  (column name: TF),  and target gene  (column name: target),
p-values and estimates of interaction}

\item{dnam}{DNA methylation matrix  (columns: samples same order as met, rows: regions/probes)}

\item{exp}{gene expression matrix (columns: samples same order as met, rows: genes)}

\item{metadata}{A data frame with samples as rownames and one columns that will be used to
color the samples}
}
\value{
A ggplot object, includes a table with results from fitting stratified model,
and the following scatter plots: 1) TF vs DNAm, 2) Target vs DNAm,
3) Target vs TF, 4) Target vs TF for samples in Q1 and Q4 for DNA methylation,
5) Target vs DNAm for samples in Q1 and Q4 for the TF
}
\description{
Create several plots to show interaction data
TF expression with target gene interaction using a linear model
\deqn{log2(RNA target) ~ log2(TF)}
to samples with highest DNAm values (top 25 percent) and lowest DNAm values (bottom 25 percent), separately.
}
\examples{
library(dplyr)
dnam <- runif(20,min = 0,max = 1) \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(dnam) <- c("chr3:203727581-203728580")
colnames(dnam) <- paste0("Samples",1:20)

exp.target <-  runif(20,min = 0,max = 10) \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(exp.target) <- c("ENSG00000232886")
colnames(exp.target) <- paste0("Samples",1:20)

exp.tf <- runif(20,min = 0,max = 10) \%>\%
  matrix(ncol = 1) \%>\%  t
rownames(exp.tf) <- c("ENSG00000232888")
colnames(exp.tf) <- paste0("Samples",1:20)

exp <- rbind(exp.tf, exp.target)

triplet <- data.frame(
   "regionID" =  c("chr3:203727581-203728580"),
   "target" = "ENSG00000232886",
   "TF" = "ENSG00000232888"
)

results <- stratified_model(triplet = triplet,dnam = dnam, exp = exp)
plots <- plot_stratified_model(
    triplet.results = results,
    dnam = dnam,
    exp = exp
)
\dontrun{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21.log2")
triplet <- data.frame(
   "regionID" = rownames(dna.met.chr21)[1:5],
   "TF" = rownames(gene.exp.chr21.log2)[11:15],
   "target" = rownames(gene.exp.chr21.log2)[1:5]
)
results <- stratified_model(
   triplet = triplet,
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2
)
plots <- plot_stratified_model(
   triplet.results = results[1,],
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2
)
# Adding color to samples
metadata <- clinical[,"sample_type",drop = FALSE]
plots <- plot_stratified_model(
   triplet.results = results[1,],
   dnam = dna.met.chr21,
   exp = gene.exp.chr21.log2,
   metadata = metadata
)
}
}
