% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap2022.R
\name{readRemap2022}
\alias{readRemap2022}
\title{Access REMAP2022 non-redundant peaks}
\usage{
readRemap2022(cell_line)
}
\arguments{
\item{cell_line}{filter peaks using cell line description field}
}
\description{
Access REMAP2022 non-redundant peaks
}
