% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_analysis.R
\name{stratified_model}
\alias{stratified_model}
\title{Fits a stratified linear model to triplet data (Target, TF, DNAm)}
\usage{
stratified_model(triplet, dnam, exp)
}
\arguments{
\item{triplet}{Dataframe with columns for region (regionID), TF  (TF), and target gene  (target),}

\item{dnam}{DNA methylation matrix  (columns: samples in the same order as met, rows: regions/probes)}

\item{exp}{A log2 (gene expression + 1) matrix with samples as columns in the same order as met
and genes as rows represented by ensembl IDs ENSG00000239415)}
}
\value{
A dataframe with Region, TF, Estimates and P-value from linear model
}
\description{
The model with interaction helps to identify DNAm changes that work synergistically with TFs in regulating
target gene expression.
}
\details{
This function fits linear model
\code{log2(RNA target) ~ log2(TF)}

To account for confounding effects from covariate variables, first use the \code{get_residuals} function to obtain
RNA residual values which have covariate effects removed, then fit interaction model. Note that no
log2 transformation is needed when \code{interaction_model} is applied to residuals data.
\code{RNA target residuals ~ TF residuals}
#'
}
\examples{
data("dna.met.chr21")
dna.met.chr21 <- map_probes_to_regions(dna.met.chr21)
data("gene.exp.chr21")
triplet <- data.frame("regionID" = rownames(dna.met.chr21)[1:10],
                      "TF" = rownames(gene.exp.chr21)[11:20],
                      "target" = rownames(gene.exp.chr21)[1:10])
results <- stratified_model(triplet, dna.met.chr21, gene.exp.chr21)
}
